requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==4.1.1
djangorestframework==3.13.1
django-rest-knox==4.2.0
django-allauth==0.52.0
drf-spectacular==0.21.2
pywatchman==1.4.1
GDAL>=3.3.2
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
django-cors-headers==3.13.0
